38G1FH7 lift the two popovers out of the control row
The series picker and the scope picker are self-contained widgets — a trigger,
a popover, a list, and the dismissal hook both share — and they took up more of
the file than the row that places them. Split so the row is about the row.

The button styles follow whichever side wears them; mutedStyle is worn on both.
